+++ packages/components/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 3.2.0
+### Added
+- Callback for table sorting added
+- pageSize prop added to pagination
+
 ### 3.1.0
 ### Changed
 - Storybook Addons Info replaced by Addon Docs